:truck: chore: organize use-cases' files by entity
