+++ api/src/core/use-cases/associates/fetch-associate.ts
