+++ api/src/core/use-cases/associates/register-associate.ts
