+++ api/src/core/use-cases/payments/register-payment.ts
